requirements.txt
requests
psycopg2-binary
python-dotenv
apache-airflow==2.9.1